Temp

+++ app/src/main/java/com/example/words/other/QuizResultFeedbackHelper.kt
@@ -5,9 +5,9 @@ import com.example.words.models.QuizResult
 class QuizResultFeedbackHelper {
     companion object {
         fun feedback(result: QuizResult): String {
-            val percentage = result.correctAnswers / result.numberOfQuizzes
+            val percentage = (result.correctAnswers.toDouble() / result.numberOfQuizzes.toDouble())
 
-            if (percentage == 1) {
+            if (percentage == 1.0) {
                 return "You are amazing!"
             } else if (percentage >= 0.5) {
                 return "You are good!"
